Add output_csv command to dbtool. (#17)
This is just the node version of a PSQL command I ran after scraping last year's tax bills:

```
\copy (select bbl, success, info->'soa'->0->>'rentStabilizedUnits' as rent_stabilized_units from pluto_19v1_2019_soa) To 'boop.csv' With CSV;
```

however, this node implementation is much more complicated and potentially runs slower.  If I knew how nontrivial it'd be to implement this, I would've just added instructions on doing it the PSQL way in the readme, but I guess it's done now so might as well merge it.