handle corrupted pdf downloads (#41)
adds check that PDF download was not corrupted (resulting in empty txt file after conversion). If it is, cached pdf and txt files are deleted and an error is thrown. These errors are logged in the DB table, and can be cleared and re-scraped just like other errors. This should solve the problem we ran into this year that some unknown number of PDF were corrupted and the scrape results look identical to buildings with valid PDFs but simply no registered RS units.