fix dockerfile hanging on tzdata step (#42)
When building the docker container there were a couple new issues breaking things:

There are some debian security files that couldn't be found, and this was solved by this SO answer updating source lists so it looks in the correct place where these files now live
Then it was getting hung up on tzdata when installing packages. This medium article explains that there is some issue with it needing a timezone and it just waits for you to interactively enter it, but we can just specify it upfront.