remove apt-get upgrade from dockerfile (#43)
For some reason the docker container fails to build on apt-get upgrade only on the EC2 instance. It ttrys to install and set up python3.5 and says it's not configured. I started looking into how to do this, but I don't think we actually need to run this upgrade, so just removed it and everything built and runs fine.