use fee id to avoid double counting RS units (#51)
In some newer formats of tax bill documents the same fee will appear multiple times in a document (eg. when there is a "previous year" or "future activity" section) which led to us double counting rent stabilized units in some cases. So now we use the unique 8-digit "fee identifier" that appears on each RS fee line to avoid counting duplicates.

This PR updates the rent stab unit regex extract function to use the fee ID, adds a duplicate fee line to the sample data in the existing test, and also completes the update of PDF-to-TXT version to 4.05 (6fe2310).

Example of a document with duplicate fees: https://nyc-doffer.s3.amazonaws.com/pdf/2/04667/0074/soa-2024-06-01.pdf.br

The regex now requires there to be a fee identifier, but after looking up many examples over the years I don't see any cases where it is absent. After we do the scrape for this year though we should do some additional checking of any properties that had RS units in the past but have 0 this year just to make sure this isn't causing a new problem if the identifier is ever missing.

[sc-15955]