handle SOA link table parsing format change (#52)
Starting midway through our 2025 scrape DOF made a small formatting change to the table listing all the SOA pdf links, and our parsing was silently failing to capture the links leaving the PDF not downloaded and results empty.

The quarter for the SOA was moved over to a previously empty cell.
image
image

Since we cache the html page containing this table and we want to be able to rerun the process on past years our fix needs to handle both table formats.

With this fix in we can restart the 2025 scrape and at least we can skip the web scraping process of submitting the search etc and can get straight to the pdf downloading/parsing - so hopefully it won't take too much extra time